backend/requirements.txt
# Web framework
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# Google Drive
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.120.0

# Document parsing
PyPDF2>=3.0.0
python-docx>=1.1.0

# Embeddings
google-genai>=1.0.0

# Vector store
faiss-cpu>=1.8.0

# LLM (optional)
openai>=1.12.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.26.0
pydantic>=2.6.0
numpy>=1.26.0
